requirements.txt
discord.py==2.3.2
openai==1.51.0
feedparser==6.0.11
apscheduler==3.10.4
supabase==2.9.1
python-dotenv==1.0.1
httpx==0.27.2
pytz==2024.2
